Split code out into functions
Previously, the code was all lumped together in one function.
That makes sense for a rough draft, but in order to refine
the joke-generation algorithm without lots of duplicated code,
it is helpful to be able to distinguish said algorithm from
the rest of the code. This will also make it easier to
extend the joke-telling capabilities in the future.