Add vague note on the limitations of the parser
The parser currently has a variety of limitations. This commit
adds a small note to the code, but is mainly here so that the
limitations are noted in the commit log itself.

The parser currently assumes that the definition will always
be the thing in <li> tags. Actually, sometimes there are no <li>
tags, so in those cases, the parser will error and no joke will
be told. :(

The parser attempts to get the 'body' of the definition, excluding
the word to be defined (as that would spoil the joke). The
implementation is fragile and unreliable; after finding a thing in
<li> tags on the page with the definition, it splits the data
within those tags on an '(n)', takes the second thing from the split
data, and strips off a number of characters matching the length
of the word to be defined. This is because the word to be defined
comes immediately after the 'n'.

However, 'n' indicates nouns, so this does not always work.
Also, sometimes the word to be defined appears elsewhere
in the definition, in usage examples, so in the
future those should also be split out.

This implementation should probably be completely rewritten so
that it looks for more reliable formatting cues.