Improve meaning-gatherer, adjust sleeps and fix newlines
This commit improves the meaning-scraper, so that it only gets the
meaning of the words from the dictionary. Previously, the meaning-
scraper would often get the word itself by mistake, or would cut
off part of the definition. This commit adjusts the code to only
take content between the second pair of brackets in the
description of the word, avoiding the word itself and ensuring
a full definition.

This commit also increases the duration of the sleep
between displaying the meaning and the punchline, allowing the
reader to attempt to guess the joke.

Finally, this commit makes the punchline all display on one line,
as a non-functional 'for' loop to strip newlines from the ends
of words has been replaced by some code that works.