Reduce usage of global variables
In some functions, this code used global variables rather than
the variables passed as parameters to the functions, as variable
names had not been updated to match the parameters when the
code was restructured. The order of function definitions meant
that these did not raise exceptions,
so were missed until the code was used in a different context,
in which it would fail to execute.

This commit fixes these variables to refer to the function
parameters.